pyproject.toml
[project]
name = "fastapi_auth_middleware"
authors = [
    { name = "Jonas Scholl", email = "jonas@code-specialist.com" },
    { name = "Yannic Schröer", email = "yannic@code-specialist.com" }
]
maintainers = [
    { name = "Jonas Scholl", email = "jonas@code-specialist.com" },
    { name = "Yannic Schröer", email = "yannic@code-specialist.com" }
]
readme = "README.md"
keywords = ['FastAPI', 'Auth', 'Middleware', 'Authentication', 'Authorization']
classifiers = [
    'Development Status :: 3 - Alpha',
    'Intended Audience :: Developers',
    'Topic :: Internet :: WWW/HTTP :: Session',
    'Topic :: Internet :: WWW/HTTP :: WSGI',
    'Topic :: Software Development :: Libraries :: Application Frameworks',
    'Topic :: Software Development :: Libraries :: Python Modules',
    'Framework :: FastAPI',
    'License :: OSI Approved :: Apache Software License',
    'Programming Language :: Python :: 3.8',
]
requires-python = ">=3.8"
dynamic = ["version", "description"]
dependencies = [
    "anyio>=3.5.0",
    "fastapi>=0.73.0",
    "idna>=3.3",
    "pydantic>=1.9.0",
    "sniffio>=1.2.0",
    "starlette>=0.17.1",
    "typing_extensions>=4.0.1",
    "python-jose>=3.3.0"
]

[project.urls]
Documentation = "https://github.com/code-specialist/fastapi-auth-middleware"
Source = "https://github.com/code-specialist/fastapi-auth-middleware"